Core: Move docgen JSDoc parser into csf-tools
