+++ code/core/src/csf-tools/index.ts
@@ -1,6 +1,7 @@
 export * from './CsfFile.ts';
 export * from './ConfigFile.ts';
 export * from './getStorySortParameter.ts';
+export * from './jsdoc.ts';
 export * from './enrichCsf.ts';
 export { babelParse } from 'storybook/internal/babel';
 export { vitestTransform } from './vitest-plugin/transformer.ts';
